vmops: Hide traceback in handled exception
Using log.exception also prints traceback, but if this exception
is handled and traceback is saved in instance, then there's no need
to print that in logs. I've came across this when trying to remove
snapshots from 5 nodes at once (in ha deployer), but some of those
nodes does not exist.